🔧(terraform) ignore the kubernetes version attribute
When declaring a kubernetes cluster in terraform with the OVH
provider, we can optionnaly specify the kubernetes version to install.
By default, the latest version is installed.
If the version is changed in terraform, the whole k8s cluster and
associated resources are destroyed and re-created. This is dangerous.
And when the version is not specified, terraform wants to recreate it
each time we apply the configuration.
Since kubernetes upgrades will be handled out of terraform (via the
OVH upgrade mechanism), we want to ignore the `version` attribute.