✨(k8s) add templates to deploy jibri on kubernetes
We want to provide a simple way to deploy jibri on kubernetes.

Kustomize (kustomize.io) is one of the most popular templating system
for Kubernetes. Among all its features, it allows to create ConfigMap
and Secrets from env files.

The templates are provided as examples, to have a ready-to-use
solution. They can also be used as a base Kustomize project, to be
easily extended and customized.