✨(k8s) autoscale jibri deployment with HPA
We want to autoscale the number of replicas of the jibri deployment
based on its usage. We want to always have 2 jibri pods available to
handle recordings. To achieve that, we declare a HPA (Horizontal
Pod Autoscaler) based on custom metrics.

The prometheus adapter configuration has been updated to add expose
jibri prometheus metrics as k8s custom metrics.