✨(k8s) terminate jibri pods gracefully to avoid service interruption
When a deployment is scaled down, we cannot chose wisely which pod we
want to terminate. So, when HPA scales down jibri deployment, it might
choose to terminate an active jibri instance.

To overcome this problem, we initiate a gracefull shutdown of jibri
when a pod is marked for deletion. And we give it a very long
terminationGracePeriodSeconds to let jibri finish its current
recording session.