✨(k8s) add side container to update the pod deletion cost
When we scale down a jibri deployment, we want to terminate pods that
are idle instead of interrupting an active jibri recording session.
Kubernetes 1.21 introduced an alpha feature named PodDeletionCost. It
allows to set a preference regarding which pods to remove first when
downscaling a ReplicaSet, based on an annotation.

This feature is disabled by default and must be enabled with the
feature gate.

This commit introduce a sidecar container that updates jibri pods
annotations when their status change.