✨(k8s) avoid terminating busy jibri pods when scaling down
The strategy that we previously implemented to avoid terminating busy
jibri instances (graceful shutdown + very long graceTerminationPeriod)
is a failure. Systematically, Pods are stuck in the Terminating state
until the graceTerminationPeriod expires. And the resources are not
released until the grace period timeout.

This commit introduce a new strategy to manage the "scale down"
problem. A sidecar container is updating jibri's pod "status" label,
with its current state (busy, idle, unknown).

The deployment selector is ignoring pods that have the label
status=busy. These pods become orphaned as long as they are busy. And
when they become idle again, they are automatically reattached to
their replicaset (and consequently, to their deployment).

As a result, when the deployment is scaled down, only pods with a
status idle or unknown can be terminated. The only risk left is if a
scale down is triggered at the beginning of a recording session, if
the status label is not yet updated, kubernetes might terminate it.