⚡️(k8s) improve and simplify jibri HPA
With jibri HPA, we want to ensure that there is at least :
 - a specific count of jibri pods available (TARGET_MIN_VALUE)
 - a specific percentage of jibri pods available across all jibri pods
   (TARGET_PERCENT)

A metric rule was specified for each of these requirements.

The first one (TARGET_MIN_VALUE) is not necessary anymore, since the
jibri Deployment is only managing jibri pods that are not busy. And we
can guarantee that this requirement is met with the "minReplicas"
instruction.

This is much easier to understand like this, because the former rule
was based on a metric that was quite complex. Also, it is more stable
this way, because HPA can instantiate a slightly larger number of pods
than necessary, while it adjusts.