✨(k8s) use separate nodepool for jibri
At this time, we create a single nodepool on which we schedule all our
pods. Some of them are related to jibri, while others are related to
prometheus or, in a near future, other jitsi components.

Some pods can prevent the cluster autoscaler to remove a jibri node,
for example if they use local storage. (see:
https://github.com/kubernetes/autoscaler/blob/master/cluster-autoscaler/FAQ.md#what-types-of-pods-can-prevent-ca-from-removing-a-node)

To avoid wasting resources while using the cluster autoscaler feature,
we want to isolate jibri pods on a specific nodepool, with autoscaling
enabled. And use another pool for resources that don't need
autoscaling.