✨(terraform) install nginx ingress controller
We want to be able to serve HTTP(S) traffic using Kubernetes
ingresses. By default, on Scaleway Kapsule, no ingress controller is
installed on the Kubernetes Cluster. We can either install one with
their web console, or manually.

The nginx-ingress controller have been installed with Helm. By
default, the ingress service uses a volatile IP address. That's why we
reserved a LoadBalancer IP on Scaleway and bound it to the service.