✨(k8s) add jvb component
We want to deploy Jitsi videobridge[1] on our kubernetes cluster. We
want it to run on a dedicate nodepool with cluster autoscaling
enabled.

We use the docker image distributed by jitsi[2].

The settings are stored in .env file used by Kustomize to generate a
ConfigMap. All values can be overriden in a kustomize overlay, if
necessary.

In its current state, no Horizontal Pod Autoscaler is configured for
JVB, so it should be scaled manually (but k8s nodes are still
added/removed accordingly).

[1](https://github.com/jitsi/jitsi-videobridge)
[2](https://github.com/jitsi/docker-jitsi-meet)