🩹(overlays) improve templates
- Add new variables in templates to prevent errors when creating
new overlays.
- Change default octo bridge selection strategy to SingleBridge.
SplitBridge is used for testing and will cause the conference to
crash if the number of JVB is less than the number of attendees.
SingleBridge will use only one bridge, it is safer to set the
default value to SingleBridge as the default number of JVB is 1
(there is not autoscaling).