:zap:(k8s) dynamic jvb pod deletion cost
We need to have a dynamic pod deletion cost for JVB in order
to kill the JVB that have less participants when scaling down.

It is useless to set the pod-deletion-cost updater
interval period with a smaller value than 15s because
it is the HPA interval to operate.
In order to limit call to k8s API, it should be set to
a much higher value, so it was set here to 60s