✨(octo) add choice for octo strategy
Octo strategy can be chosen among three options, which are:
- SingleBridge: conference is binded to one jvb, can not move from it
- SplitBridge: conferences will be split between jvbs, for dev
- RegionBased: conferences are managed based on regions,
and are binded to one jvb until it is overloaded
(overloaded means stressLevel > BRIDGE_STRESS_THRESHOLD)
then it will choose other jvbs in same region

Another option (IntraRegionBased) is possible,
but it was not explored enough to add it.
More precisions here:
"https://github.com/jitsi/jicofo/
tree/master/src/main/java/org/jitsi/jicofo/bridge"

Creation of overlay was not done at the right place,
this was fixed.
Creation of folder is done after asking for variables
so that when people put wrong variables,
they can exit and re-try right away,
instead of having an empty folder to delete