🚸(simulation) add env file and makefile command
We need to launch simulation more easily, without
having to modify production code. A docker solution
for this solution is not effortlessly realizable
given that the script provides graphical solutions.
A makefile command has been made to facilitate the
execution of the script (which must be launched from
the root of the repository).