🔧(k8s) configure NGINX_RESOLVER to use coredns
Since version stable-7830 of jitsi docker, nginx is not able anymore to
resolve url used internally by services. The service prosody.jitsi.svc
for example is not resolved and nginx is not able to fetch prosody. To
resolve this issue, we choose to ser the variable NGINX_RESOLVER with
the coredns FQDN service url and then use FQDN everywhere in the
configuration when we want to use a service url.